Move sorting utils into shared folder
